Adjusted problem three, line 21, from int.Parse to double.Parse for accuracy

+++ LinqPracticeThree/Program.cs
@@ -18,7 +18,7 @@ namespace LinqPracticeThree
                 "98,100,66,74,55"
             };
 
-            var gradeResults = classGrades.Select(x => x.Split(',')).Select(x => Array.ConvertAll(x, int.Parse)).Select(x => (x.Sum() - x.Min()) / (x.Count() - 1)).Average();
+            var gradeResults = classGrades.Select(x => x.Split(',')).Select(x => Array.ConvertAll(x, double.Parse)).Select(x => (x.Sum() - x.Min()) / (x.Count() - 1)).Average();
 
             Console.WriteLine(gradeResults);
             Console.ReadLine();
